Remove the usage of down and disabled together
This patch removes the trigger with state '== down || == disabled'
because it is not working. We replaced it by a check that validates the
fact that alarm is triggered only if one service is up and at least
another service is down or disabled by using the percentage of services
up.

Change-Id: Ic4f2e6ea7be9e90a59116b46e3487074a365c8bd