Allow overriding the  scheduler_host_manager
Deployments with Ironic need a custom host manager. Allow overriding it.

Don't add PciPassthroughFilter explicitly to the active filters.
nova.scheduler.filters.all_filters should include this if it's in the
default filters, and if it is removed from scheduler_default_filters in
the pillar, the new scheduler_custom_filters will allow users to
manually add it.